+++ internal/cmd/destroy.go
@@ -1,9 +1,16 @@
 package cmd
 
 import (
-	"github.com/spf13/cobra"
+	"fmt"
+	"os"
+	"os/exec"
 
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+
+	"github.com/craftcms/nitro/config"
 	"github.com/craftcms/nitro/internal/nitro"
+	"github.com/craftcms/nitro/internal/sudo"
 )
 
 var destroyCommand = &cobra.Command{
@@ -15,11 +22,49 @@ var destroyCommand = &cobra.Command{
 			machine = flagMachineName
 		}
 
+		// get the sites
+		var cfg config.Config
+		if err := viper.Unmarshal(&cfg); err != nil {
+			return err
+		}
+
+		var domains []string
+		for _, site := range cfg.Sites {
+			domains = append(domains, site.Hostname)
+		}
+
 		destroyAction, err := nitro.Destroy(machine)
 		if err != nil {
 			return err
 		}
 
-		return nitro.Run(nitro.NewMultipassRunner("multipass"), []nitro.Action{*destroyAction})
+		if err := nitro.Run(nitro.NewMultipassRunner("multipass"), []nitro.Action{*destroyAction}); err != nil {
+			return err
+		}
+
+		// delete machine config file
+		if err := os.Remove(viper.ConfigFileUsed()); err != nil {
+			// dont return the error, we still need to cleanup the hosts file
+			fmt.Println("unable to remove the config file:", err)
+		}
+
+		// prompt to remove hosts file
+		nitro, err := exec.LookPath("nitro")
+		if err != nil {
+			return err
+		}
+
+		fmt.Println("Removing sites from your hosts file")
+
+		if len(domains) == 0 {
+			return nil
+		}
+
+		cmds := []string{"hosts", "remove"}
+		for _, domain := range domains {
+			cmds = append(cmds, domain)
+		}
+
+		return sudo.RunCommand(nitro, machine, cmds...)
 	},
 }